add ng-add and move utilities
