+++ modules/entity/package.json
@@ -20,5 +20,6 @@
   "ng-update": {
     "packageGroup": "NG_UPDATE_PACKAGE_GROUP"
   },
+  "schematics": "./collection.json",
   "sideEffects": false
 }
