--- modules/schematics/CHANGELOG.md
@@ -1,121 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
-
-<a name="5.2.0"></a>
-# 5.2.0 (2018-03-07)
-
-
-### Bug Fixes
-
-* **Schematics:** Correct usage of upsert actions for entity blueprint (#821) ([1ffb5a9](https://github.com/ngrx/platform/commit/1ffb5a9))
-
-
-
-<a name="5.1.0"></a>
-# 5.1.0 (2018-02-13)
-
-
-### Bug Fixes
-
-* **Effects:** Make ofType operator strictFunctionTypes safe (#789) ([c8560e4](https://github.com/ngrx/platform/commit/c8560e4)), closes [#753](https://github.com/ngrx/platform/issues/753)
-* **Schematics:** Add store import to container blueprint (#763) ([a140fa9](https://github.com/ngrx/platform/commit/a140fa9)), closes [#760](https://github.com/ngrx/platform/issues/760)
-* **Schematics:** Remove extra braces from constructor for container blueprint (#791) ([945bf40](https://github.com/ngrx/platform/commit/945bf40)), closes [#778](https://github.com/ngrx/platform/issues/778)
-* **Schematics:** Use correct paths for nested and grouped feature blueprint (#756) ([c219770](https://github.com/ngrx/platform/commit/c219770))
-
-
-### Features
-
-* **Schematics:** Add group option to entity blueprint (#792) ([0429276](https://github.com/ngrx/platform/commit/0429276)), closes [#779](https://github.com/ngrx/platform/issues/779)
-* **Schematics:** Add upsert methods to entity blueprint (#809) ([7acdc79](https://github.com/ngrx/platform/commit/7acdc79)), closes [#592](https://github.com/ngrx/platform/issues/592)
-
-
-
-<a name="5.0.0"></a>
-# 5.0.0 (2018-01-22)
-
-
-### Bug Fixes
-
-* **Schematics:** Add group folder after feature name folder (#737) ([317fb94](https://github.com/ngrx/platform/commit/317fb94))
-* **Schematics:** Add handling of flat option to entity blueprint ([fb8d2c6](https://github.com/ngrx/platform/commit/fb8d2c6))
-* **Schematics:** Distinguish between root and feature effect arrays when registering (#718) ([95ff6c8](https://github.com/ngrx/platform/commit/95ff6c8))
-* **Schematics:** Don't add state import if not provided (#697) ([e5c2aed](https://github.com/ngrx/platform/commit/e5c2aed))
-* **Schematics:** Make variable naming consistent for entity blueprint (#716) ([765b15a](https://github.com/ngrx/platform/commit/765b15a))
-
-
-### Features
-
-* **Schematics:** Add alias for container, store and action blueprints (#685) ([dc64ac9](https://github.com/ngrx/platform/commit/dc64ac9))
-* **Schematics:** Add alias for reducer blueprint (#684) ([ea98fb7](https://github.com/ngrx/platform/commit/ea98fb7))
-* **Schematics:** Add effect to registered effects array (#717) ([f1082fe](https://github.com/ngrx/platform/commit/f1082fe))
-* **Schematics:** Add option to group feature blueprints in respective folders (#736) ([b82c35d](https://github.com/ngrx/platform/commit/b82c35d))
-* **Schematics:** Introduce [@ngrx](https://github.com/ngrx)/schematics (#631) ([1837dba](https://github.com/ngrx/platform/commit/1837dba)), closes [#53](https://github.com/ngrx/platform/issues/53)
-
-
-
-
-<a name="5.1.0"></a>
-# 5.1.0 (2018-02-13)
-
-
-### Bug Fixes
-
-* **Effects:** Make ofType operator strictFunctionTypes safe (#789) ([c8560e4](https://github.com/ngrx/platform/commit/c8560e4)), closes [#753](https://github.com/ngrx/platform/issues/753)
-* **Schematics:** Add store import to container blueprint (#763) ([a140fa9](https://github.com/ngrx/platform/commit/a140fa9)), closes [#760](https://github.com/ngrx/platform/issues/760)
-* **Schematics:** Remove extra braces from constructor for container blueprint (#791) ([945bf40](https://github.com/ngrx/platform/commit/945bf40)), closes [#778](https://github.com/ngrx/platform/issues/778)
-* **Schematics:** Use correct paths for nested and grouped feature blueprint (#756) ([c219770](https://github.com/ngrx/platform/commit/c219770))
-
-
-### Features
-
-* **Schematics:** Add group option to entity blueprint (#792) ([0429276](https://github.com/ngrx/platform/commit/0429276)), closes [#779](https://github.com/ngrx/platform/issues/779)
-* **Schematics:** Add upsert methods to entity blueprint (#809) ([7acdc79](https://github.com/ngrx/platform/commit/7acdc79)), closes [#592](https://github.com/ngrx/platform/issues/592)
-
-
-
-<a name="5.0.0"></a>
-# 5.0.0 (2018-01-22)
-
-
-### Bug Fixes
-
-* **Schematics:** Add group folder after feature name folder (#737) ([317fb94](https://github.com/ngrx/platform/commit/317fb94))
-* **Schematics:** Add handling of flat option to entity blueprint ([fb8d2c6](https://github.com/ngrx/platform/commit/fb8d2c6))
-* **Schematics:** Distinguish between root and feature effect arrays when registering (#718) ([95ff6c8](https://github.com/ngrx/platform/commit/95ff6c8))
-* **Schematics:** Don't add state import if not provided (#697) ([e5c2aed](https://github.com/ngrx/platform/commit/e5c2aed))
-* **Schematics:** Make variable naming consistent for entity blueprint (#716) ([765b15a](https://github.com/ngrx/platform/commit/765b15a))
-
-
-### Features
-
-* **Schematics:** Add alias for container, store and action blueprints (#685) ([dc64ac9](https://github.com/ngrx/platform/commit/dc64ac9))
-* **Schematics:** Add alias for reducer blueprint (#684) ([ea98fb7](https://github.com/ngrx/platform/commit/ea98fb7))
-* **Schematics:** Add effect to registered effects array (#717) ([f1082fe](https://github.com/ngrx/platform/commit/f1082fe))
-* **Schematics:** Add option to group feature blueprints in respective folders (#736) ([b82c35d](https://github.com/ngrx/platform/commit/b82c35d))
-* **Schematics:** Introduce [@ngrx](https://github.com/ngrx)/schematics (#631) ([1837dba](https://github.com/ngrx/platform/commit/1837dba)), closes [#53](https://github.com/ngrx/platform/issues/53)
-
-
-
-
-<a name="5.0.0"></a>
-# 5.0.0 (2018-01-22)
-
-
-### Bug Fixes
-
-* **Schematics:** Add group folder after feature name folder (#737) ([317fb94](https://github.com/ngrx/platform/commit/317fb94))
-* **Schematics:** Add handling of flat option to entity blueprint ([fb8d2c6](https://github.com/ngrx/platform/commit/fb8d2c6))
-* **Schematics:** Distinguish between root and feature effect arrays when registering (#718) ([95ff6c8](https://github.com/ngrx/platform/commit/95ff6c8))
-* **Schematics:** Don't add state import if not provided (#697) ([e5c2aed](https://github.com/ngrx/platform/commit/e5c2aed))
-* **Schematics:** Make variable naming consistent for entity blueprint (#716) ([765b15a](https://github.com/ngrx/platform/commit/765b15a))
-
-
-### Features
-
-* **Schematics:** Add alias for container, store and action blueprints (#685) ([dc64ac9](https://github.com/ngrx/platform/commit/dc64ac9))
-* **Schematics:** Add alias for reducer blueprint (#684) ([ea98fb7](https://github.com/ngrx/platform/commit/ea98fb7))
-* **Schematics:** Add effect to registered effects array (#717) ([f1082fe](https://github.com/ngrx/platform/commit/f1082fe))
-* **Schematics:** Add option to group feature blueprints in respective folders (#736) ([b82c35d](https://github.com/ngrx/platform/commit/b82c35d))
-* **Schematics:** Introduce [@ngrx](https://github.com/ngrx)/schematics (#631) ([1837dba](https://github.com/ngrx/platform/commit/1837dba)), closes [#53](https://github.com/ngrx/platform/issues/53)
