--- modules/schematics/README.md
@@ -1,5 +1,0 @@
-# @ngrx/schematics
-
-The sources for this package are in the main [NgRx](https://github.com/ngrx/platform) repo. Please file issues and pull requests against that repo.
-
-License: MIT
