--- modules/schematics/collection.json
@@ -1,53 +1,0 @@
-{
-  "extends": ["@schematics/angular"],
-  "schematics": {
-    "action": {
-      "aliases": [ "a" ],
-      "factory": "./src/action",
-      "schema": "./src/action/schema.json",
-      "description": "Add store actions"
-    },
-
-    "container": {
-      "aliases": [ "co" ],
-      "factory": "./src/container",
-      "schema": "./src/container/schema.json",
-      "description": "Add store container component"
-    },
-
-    "effect": {
-      "aliases": [ "ef" ],
-      "factory": "./src/effect",
-      "schema": "./src/effect/schema.json",
-      "description": "Add side effect class"
-    },
-
-    "entity": {
-      "aliases": [ "en" ],
-      "factory": "./src/entity",
-      "schema": "./src/entity/schema.json",
-      "description": "Add entity state"
-    },
-
-    "feature": {
-      "aliases": [ "f" ],
-      "factory": "./src/feature",
-      "schema": "./src/feature/schema.json",
-      "description": "Add feature state"
-    },
-
-    "reducer": {
-      "aliases": [ "r" ],
-      "factory": "./src/reducer",
-      "schema": "./src/reducer/schema.json",
-      "description": "Add state reducer"
-    },
-
-    "store": {
-      "aliases": [ "st" ],
-      "factory": "./src/store",
-      "schema": "./src/store/schema.json",
-      "description": "Adds initial setup for state managment"
-    }
-  }
-}
