--- modules/schematics/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@ngrx/schematics",
-  "version": "5.2.0",
-  "description": "NgRx Schematics for Angular",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ngrx/platform.git"
-  },
-  "keywords": [
-    "RxJS",
-    "Angular",
-    "Redux",
-    "NgRx",
-    "Schematics",
-    "Angular CLI"
-  ],
-  "author": "Brandon Roberts <robertsbt@gmail.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/ngrx/platform/issues"
-  },
-  "homepage": "https://github.com/ngrx/platform#readme",
-  "schematics": "./collection.json",
-  "peerDependencies": {
-    "@angular-devkit/core": "^0.5.0",
-    "@angular-devkit/schematics": "^0.5.0"
-  }
-}
